backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
anthropic==0.44.0
httpx==0.28.1
playwright==1.49.1
celery[redis]==5.4.0
redis==5.2.1
python-dotenv==1.0.1
python-multipart==0.0.20
pillow==11.1.0
stripe==11.4.1
beautifulsoup4==4.12.3
pyjwt==2.10.1
bcrypt==4.2.1
celery[redis]==5.4.0
